Project restructure: - Moved code into layered layout:   * src/producers/ (AVD producer + example template)   * src/consumers/ (Kafka->Mongo consumer)   * src/lib/ (shared kafka_producer + spotify_payloads)   * src/app/ (Streamlit dashboard) - Updated Makefile to module-style runners:   * make run -> python -m producers.avd_producer   * make consume -> python -m consumers.kafka_consumer   * make app -> streamlit run src/app/streamlit_app.py - Kept kafka wait+init logic; topics created idempotently - Updated SETUP.md/STATUS.md to reflect new structure - No functional changes to my (AVD) data flow; paths only
